+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to Mara 3 are documented here.
 
-## [3.0.16] - unreleased
+## [3.0.16] - 2026-07-03
 
 ### Changed